Add CUDA stream management and telemetry for FDM
- Implemented CUDA stream lifecycle management in streams.cu, including creation, destruction, and synchronization of compute streams and events.
- Added telemetry functionality in telemetry.cu to gather and report energy and torque statistics during FDM steps, supporting both double and single precision.

+++ native/backends/fem/cpu/mfem/interactions/exchange_legacy_gpu_upload.cpp
@@ -36,8 +36,8 @@ bool upload_legacy_sparse_exchange_to_gpu_state(
         error = "legacy sparse exchange CSR has invalid dimensions";
         return false;
     }
-    if (height > static_cast<int>(std::numeric_limits<uint32_t>::max()) ||
-        width > static_cast<int>(std::numeric_limits<uint32_t>::max())) {
+    if (static_cast<uint64_t>(height) > std::numeric_limits<uint32_t>::max() ||
+        static_cast<uint64_t>(width) > std::numeric_limits<uint32_t>::max()) {
         error = "legacy sparse exchange CSR dimensions exceed u32 GPU indexing";
         return false;
     }
